refactor: improve C challenge solutions with input validation and docs
Add doc comments, secure scanf usage with width specifiers and return checks, cast time() for srand, simplify stdlib challenge to print random number, and fix LaTeX math delimiters in notes. Also ignore Excalidraw and .obsidian directories.

+++ Year_1/Semester_2/C_PROGRAMMING_II/Assignments/File_Handling/src/exercise5.c
@@ -17,7 +17,7 @@ int main() {
         return -1;
     }
     for (int i = 1; i < 6; i++){
-        size_t written = fprintf(file, "%d\n", i);
+        int written = fprintf(file, "%d\n", i);
         if (written < 0){
             printf("error writing to file");
             fclose(file);
